Prevented rejected proposals remaining open
Without the else-if clause added here, there is nothing to prevent a proposal from being left in a pending state if it has been rejected by a majority of members. Being left in a pending state when the voting majority have rejected it could allow future members to vote to accept a proposal that should already have been rejected.